feat: migrate server.ts to typescript
